--prefix and --match
This commit adds the --prefix option for searching for all urls that
begin with a particular prefix. It uses the IA CDX API behind the
scenes. The --regex option was added to filter only particular urls
when looking them up. So for example if you are only interested in
tweets you could do something like this, which would elminate images
and other media files that hang off the same URL prefix:

    waybackprov --prefix --match '/status/\d+$' https://twitter.com/edsu/